RL_TPO/requirements.txt
torch>=2.1.0
ollama>=0.3.0
pyyaml
tqdm
numpy
pandas
matplotlib
seaborn
wandb
trl>=0.9.0
peft
